Сохраняет фильмы в локальное хранилище и показывает их в офлайне

+++ src/api/provider.js
@@ -1,3 +1,13 @@
+import FilmModel from "../model/films";
+
+const createStoreStructure = (items) => {
+  return items.reduce((acc, current) => {
+    return Object.assign({}, acc, {
+      [current.id]: current
+    });
+  }, {});
+};
+
 export default class Provider {
   constructor(api, store) {
     this._api = api;
@@ -5,26 +15,51 @@ export default class Provider {
   }
 
   getFilms() {
-    this._api.getFilms();
+    if (this._isOnline()) {
+      return this._api.getFilms()
+        .then((films) => {
+          const items = createStoreStructure(films.map(FilmModel.adaptToServer));
+          this._store.setItems(items);
+          return films;
+        });
+    }
+
+    const storeFilms = Object.values(this._store.getItems());
+
+    return Promise.resolve(storeFilms.map(FilmModel.adaptToClient));
   }
 
   getComments(filmId) {
-    this._api.getComments(filmId);
+    if (this._isOnline()) {
+      return this._api.getComments(filmId);
+    }
   }
 
   updateFilm(film) {
-    this._api.updateFilm(film);
+    if (this._isOnline()) {
+      return this._api.updateFilm(film);
+    }
   }
 
   addComment(data) {
-    this._api.addComment(data);
+    if (this._isOnline()) {
+      return this._api.addComment(data);
+    }
   }
 
   deleteComment(commentId) {
-    this._api.deleteComment(commentId);
+    if (this._isOnline()) {
+      return this._api.deleteComment(commentId);
+    }
   }
 
   sync() {
-    this._api.sync();
+    if (this._isOnline()) {
+      return this._api.sync();
+    }
+  }
+
+  _isOnline() {
+    return window.navigator.onLine;
   }
 }
